Refactor signature handler and stats
